LayoutTests/inspector/console-format-collections-expected.txt

CONSOLE MESSAGE: line 14: [object NodeList]
CONSOLE MESSAGE: line 18: [object HTMLCollection]
CONSOLE MESSAGE: line 22: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 26: [object HTMLAllCollection]
CONSOLE MESSAGE: line 30: [object HTMLCollection]
CONSOLE MESSAGE: line 34: [object NodeList]
Tests that console nicely formats HTML Collections and NodeLists.

console-format-collections.html:14[<select id="sel" name="sel">]
console-format-collections.html:18[<script src="evaluate-in-frontend.js">, <script>]
console-format-collections.html:22[<option value="1">one</option>, <option value="2">two</option>]
console-format-collections.html:26[<html>, <head>, <script src="evaluate-in-frontend.js">, <script>, <body onload="onload()">, <p> Tests that console nicely formats HTML Collections and NodeLists. </p>, <div style="display:none">, <form id="f">, <select id="sel" name="sel">, <option value="1">one</option>, <option value="2">two</option>, <input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">, <div id="frontend-script" style="display:none">, <div id="output">]
console-format-collections.html:30[<select id="sel" name="sel">, <input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">]
console-format-collections.html:34[<input type="radio" name="x" value="x1">, <input type="radio" name="x" value="x2">]
